Updated overloaded operator =(Image&& img) Added template at function Began updating code to make use of at Started switching unsigned char/short to uint8/16_t Changed Convert member function Updated ImRead function

+++ ImageStacking.cpp
@@ -55,7 +55,7 @@ void IS::Average(std::vector<Image>& imgvec,Image& final_image){
     
 void IS::Median(std::vector<Image>& imgvec, Image& final_image) {
     float* fptr = (float*)final_image.data;
-    std::vector<ushort> pixelstack(imgvec.size());
+    std::vector<float> pixelstack(imgvec.size());
 
 #pragma omp parallel for firstprivate(pixelstack)
     for (int el = 0; el < final_image.rows * final_image.cols; ++el) {
